// src/types/Discover.ts
export interface DiscoverResponse {
	page?: number;
	results?: DiscoverResult[];
	total_results?: number;
	total_pages?: number;
}

export interface DiscoverResult {
	poster_path?: string | null;
	adult?: boolean;
	overview?: string;
	release_date?: string;
	genre_ids?: number[];
	id?: number;
	original_title?: string;
	original_language?: string;
	title?: string;
	backdrop_path?: string | null;
	popularity?: number;
	vote_count?: number;
	video?: boolean;
	vote_average?: number;
}